Fixed invalid argument exception in error message. 	Change on 2012/11/29 by tball@tball ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=39433357

+++ src/main/java/com/google/devtools/j2objc/util/UnicodeUtils.java
@@ -66,7 +66,7 @@ public final class UnicodeUtils {
             if (!isValidCppCharacter(value)) {
               if (logErrorMessage) {
                 J2ObjC.error(String.format("Illegal C/C++ Unicode character \\u%4x in \"%s\"",
-                    value, s));
+                    (int) value, s));
               } else {
                 J2ObjC.error();
               }
